upload: Support supplied image for an image transfer.
For REST API support, we want to be able to execute the upload
command with a supplied image. Currently when executing the command,
the image is created. This patch will create the image only if no
image was supplied, otherwise, set needed parameters for the command
according to the supplied image.

Add tests for the upload command, verifying this change.

Change-Id: I41066d1b8d2d7b18e4ceb12e1754b99189b3b258
Bug-Url: https://bugzilla.redhat.com/1337077
Signed-off-by: Amit Aviram <aaviram@redhat.com>