core: set display/migration for cluster network
The display and migration flag should be set for the default cluster
network.

The upgrade script that did that before in 03_05 and was squashed to the
database used the name 'ManagementNetwork' to recognize this network.
However, the name were changed to 'DefaultManagementNetwork' and this
change is now running in the pre-upgrade step, so those flags update
never occurred.

This patch sets the correct value for display and migration flags for
the default cluster network.

Change-Id: Iae220e05b8908ea372c134db12f2c0fa5466b7de
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1372955
Signed-off-by: emesika <emesika@redhat.com>