core: CommandBase - CommandStep/FINALIZING step
This patch prevents adding a finalizing step to steps
added by overriding getCommandStep().

The finalizing step is currently used for steps and not for jobs only in
very specific cases in which also validation and execution steps are
added to a step/sub step.

When command is adding a step by overriding getCommandStep() its intended
to add a "simple" step as replacement to the async tasks added steps when
jobs are used instead of tasks therefore there's no need to add a finalizing
step when the step was added by getCommandStep().

Change-Id: I9c6d7a5aebfbfa176fca5e8d90b6127ee7cc7a5f
Signed-off-by: Liron Aravot <laravot@redhat.com>