webadmin: infinite loop in new VM/pool dialog
Open new VM/pool dialog and pick a template which had non-null custom
compatibility version, the dialog ends up in infinite refresh loop.

The reason was that the CoreVmBaseToUnitBuilder.postBuild called the
model.getCustomCompatibilityVersion().setSelectedItem(vm.getCustomCompatibilityVersion());
all the time regardless of the fact it changed or not. This change leads to a
full refresh of the whole dialog, including this builder.

Fixed by setting this value only if changed.

Change-Id: I6c2e24faf1793f9de0e973ee41425f7cd32ea922
Bug-Url: https://bugzilla.redhat.com/1373204
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>