core: extract disk statistics from vm-dynamic
This is quite an absurd that VmDynamic#disks was set with
@UnchangeableByVdsm although it is reported only by VDSM.
This might be an indication of a problem with the design
- it represents disk statistics and thus should not be
part of VmDynamic:

1. It is not supposed to be compared with value from the
database (hence why it is set with @UnchangeableByVdsm)

2. It is not supposed to be sent for each VM that is being
queried, including search queries.

Therefore, this patch extracts this data from VmDynamic to
VdsmVm that contains similar type of statistics.

Change-Id: I4139db014d4663c1ee1eff5037939db628b2f53e
Signed-off-by: Arik Hadas <ahadas@redhat.com>