core: Validate images existance for unregistered Template.
Once an unregister Template entity is being registered to the engine,
the engine should verify all its disks existance in the Storage.
The reason the engine should perform this validation is since the
Template's OVF might not be synced with the disks in the Storage.

If one of the disks doesn't exists, the engine should block
the operation.
This behavior should be compatible with the behavior being performed
today once a Template is being imported from an export domain.

Change-Id: I3b890bd831c34770d6fd43416d5b15888b45f85b
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1367488