restapi: Validate 'nfs_timeo' and 'nfs_retrans' range
Current the 'nfs_timeo' and 'nfs_retrans' values used when adding NFS
storage aren't validated, their values are just truncated and passed to
the backend. That is wrong.

As the backend receives a 'short' and the API uses an 'integer' it isn't
possible to do the validation in the backend, thus it needs to be done
in the API.

This patch changes the API so that it performs explicitly this check
before trying to do anything with the storage. If the check fails, for
example if the value of 'nfs_timeo' is 131077, the operation will fail
with error code 400 (Bad Request) and the following response body will
be returned:

  <fault>
    <reason>Value out of range</reason>
    <detail>The value 131077 of attribute 'nfs_timeio' is outside of range 0 to 65535.</detail>
  </fault>

Change-Id: I3e26b58879b2a3e0ed54fbdb9a1c70aa4efcc73c
Bug-Url: https://bugzilla.redhat.com/1373581
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>