core: QuotaManagerTest throws declarations
Allow methods to declare a thrown exception (which is never actually
thrown) instead of having to catch it in order to clean up the code.

Besides cleaning up the code, this patch has an added bonus of actually
making the test fail if the underlying implementation is ever broken in
a way that a CloneNotSupportedException is thrown.

Change-Id: Ieddb8c728258be8176b24fb97109dda79e60e47e
Signed-off-by: Allon Mureinik <amureini@redhat.com>