core: Remove DbUtils
DbUtils' closeQuietly methods were just a nicety to avoid
boiler-plated code in the finally block of a try block handling JDBC
resources. Since Java 7, all JDBC resources are autoclosable, so
there's a much cleaner way of handling these issues - with a
try-with-resource block.

This patch converts all the usages of DbUtils to standard Java
try-with-resource blocks and removes this now-useless class
completely.

Change-Id: I6620228031c9f723b350e10820c5a8035fef241f
Signed-off-by: Allon Mureinik <amureini@redhat.com>