engine: Clear IDE disk address on bootable change
In the following scenario, the VM will fail to start:
1. Create new VM with IDE disk, verify it starts
2. Add second IDE disk
3. Turn off bootable flag from 1st disk, make 2nd disk bootable
4. Start the VM

The root cause of this issue is that we save the device address
of the disk, but we changed the order of the disks.
This will cause a mismatch between the address and the name of the disk.

In order to fix this issue, the address will be cleared for IDE disks
if the 'bootable' attribute changed.

Change-Id: I46847ae9f0870e454e81722834535861b87abed8
Bug-Url: https://bugzilla.redhat.com/1186817
Signed-off-by: Fred Rolland <frolland@redhat.com>