core: Use CDI.current() to remove internal deps from Injector
Since CDI spec 1.2 programatic lookup can be achieved using the
CDI.current() which locates the CDI implementation and has a reference
to the BeanManager. Thus the injector needs not to be a part of the EJB
lifecycle and can be moved to utils.

The usage of Injector is mainly for unmanaged code. Direct usage of
CDI.current().select(Some.class, Annotations...) is preferred and more
complete and the usage of Injector can be kept only to to help with
injecting members into commands (i.e the various command factories we
have)

- Injector is no more a Bean, all inner deps are being removed
- Uses CDI.current() to do programmatic lookups
- Prefer CDI.current().select(Bean...) over Injector.get(Bean...)
- Simplified InjectorRule impl

Change-Id: Ibc60a4a38a184954f782e3ca3b5563fc55050776
Signed-off-by: Roy Golan <rgolan@redhat.com>