restapi: Build correctly creation status links
The code that builds the links returned by the creation status resource
doesn't build correctly the links, because it doesn't take into account
the source of the API version. As a result the links contain always the
version prefix. For example:

  <creation href="/ovirt-engine/api/v4/vms/..." id="...">
    <status>complete</status>
  </creation>

Note the '/v4' suffix, it should be there only if it is included in the
request URI, but it is currently always present, even when the version
is selected using the 'Version' header.

This patch fixes the creation status resource so that it builds the link
taking into account where the version has been specified.

Change-Id: I9cc0f7f03d2dd5ece00ed0081a7fb058cd5b7a27
Related-To: https://bugzilla.redhat.com/1359139
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>