core: Introduce DiskStorageTypePredicate
Introduced the DiskStorageTypePredicate class to lay the groundwork for
implementing DiskFilters for any concrete type that extends Disk.

The current patch just does the minimal work of converting the
ONLY_IMAGES predicate to use this new class, and the following patches
will introduce similar implementations for LunDisks and CinderDisks.

In order to perform this filtering, a forClass(Class<? extends Disk>)
method was added to DiskStorageType so it can be the only place that
holds the correlation between a DiskStorageType and the concrete Disk
class that matches it. Additionally, DiskStorageTypeTest was
introduced to verify that each concrete Disk type has a matching
DiskStorageType and that no DiskStorageType matches more than one Disk
class.

Change-Id: I4edf05438f426216f75792182e0174392894f9a6
Signed-off-by: Allon Mureinik <amureini@redhat.com>