webadmin: Fix incorrect VM status icon in NUMA menu
This patch fixes an error wherein the status icon of a VM could be
out of sync with the actual status of the VM in the NUMA pinning menu.

Cause:
The instance of the VM model pulled from the database was being discarded
and the version currently stored in memory was being used to back the
VmNumaSupportModel instead.

Fix:
Use the model instance being returned from the database to back the
VmNumaSupportModel instead of the current model stored in memory.

Change-Id: Idd5ed3807d90dbd0ed6b76a31d6a1673e52eb988
Bug-Url: https://bugzilla.redhat.com/1147858
Signed-off-by: Phillip Bailey <phbailey@redhat.com>