core: TestCommon @BeforeClass and @AfterClass
All the tests in the this suite start with calling
TestCommon.initialize() and they all end with calling
TestCommon.terminate().

Instead of repeating these calls, a much cleaner approach would be to
simply annotate them as such, have all the tests extend TestCommon, and
allow JUnit to do all the heavy lifting.

Change-Id: Ia362d885ddb184d35e1d3851e43349c6797a7b5f
Signed-off-by: Allon Mureinik <amureini@redhat.com>