core: SSH test exceptions
A lot of the tests in the ssh package will catch any exception and
throw a RuntimeException, thus failing the test. A much cleaner
approach would be to just throw the original exception all the way
up. This way the test will still fail, but will at least provide some
useful information on why it failed.

Change-Id: I81a2661b6d3d5d48e548b90c20b53e243f0c5d68
Signed-off-by: Allon Mureinik <amureini@redhat.com>