core: removing changedbowner.sh script
This script was relevant when the old backup utility was used and there
was a chance that the backup was done with other user than 'engine'.
In that case, if we try to restore the database we will got permissions
issues.
Now that engine-backup is used and verifies that the user that backup
the database is 'engine', this script is no longer relevant.

Change-Id: I0474ea645e2f3a0f4db20b510e2a1ef68e73b4ca
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1371501
Signed-off-by: emesika <emesika@redhat.com>