core: fix retrival of vm's pid
The process identifier (PID) of a VM is reported correctly by VDSM but
due to incorrect handling in the engine it is not set in VmDynamic#vmPid.
This patch fixes that handling so the PID is correctly saved to the
database. Further usages of this field need to be considered separately.

Change-Id: I83820d8181ee382c720510b3cb51d110cf8bee40
Signed-off-by: Arik Hadas <ahadas@redhat.com>