core: adding debug message to DWH heart beat
Adding debug messages on DWH heart beat start/end in order to
investigate delays that occur sometimes when the result is that the
update of the corresponding field in DB not occurs in the configured
rate.

This will not solve the reported bug, rather it will give us information
which part cause the delay, the scheduling mechanism or the database.

Change-Id: Ia32c559ee380e0dbb7279f94b7dd9a99c0670a2a
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1373456
Signed-off-by: emesika <emesika@redhat.com>