core: CreateGlusterVolumeCommand.getVolume()
Add a utility method getVolume() to return the volume from the
parameters instead of saving it as a data member.

This achieves two things:
First, from a design perspective, it saves the need to keep the same
object in two places.
Second, it allows for more flexibility, and lends ease to testing the
class in the following patches of the series.

Change-Id: If8c30fc9ea0822e3832b132ef3f13070a9f7f4f8
Signed-off-by: Allon Mureinik <amureini@redhat.com>