restapi: Recursively scan for service interfaces
The creation of links to collections relies on scanning of the
implementation classes, in order to find the implemented service
interfaces. But currently the code that does this scanning doesn't take
into account that some interfaces may not be implememented directly, but
by base classes. In those situations the links to collections aren't
created correctly. The 'BackendDataCenterNetworksResource' class is an
example of that. It doesn't directly implement any service interface,
only the base class does. To avoid that issue this patch changes the
scanning code so that it recursively scans the parent classes.

Change-Id: Ice711e9e94048ac695c4cd502e8beb563f6f2017
Bug-Url: https://bugzilla.redhat.com/1375668
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>