core: AddVmCommandTest use init()
Use init(), which is a "user-facing" method used to initialize the
"business" part of the command, instead of postConstruct() which is
geared to initialize CommandBase's internal structures which the test
does not, and should not, care about.

As a neat side-bonus, setting up the CommandContext could be dropped,
as the test really doesn't care about it.

Change-Id: I8002501d17a192a345c319c141763d7b976a62b0
Signed-off-by: Allon Mureinik <amureini@redhat.com>