core: Move OvfManager to bll
OvfManager is a tool used by the bll layer and adds abstraction and
exposes an API to export VM and Templates. Moving it to bll will allow
extending and enriching the information added to the OVF without
constantly expending the structures and arguments.

The following patch is adding a specialized OVF writer for hosted engine
and will use the OvfManager access to the DAL layer.

Change-Id: I4d82408f10de3e6172e8cce4044d5a8b58f36afc
Bug-Url: https://bugzilla.redhat.com/1358383
Signed-off-by: Roy Golan <rgolan@redhat.com>