core: ImportVMFromConfigurationCommandTest @Spy annotation
Spy the command with the @Spy annotation instead of explicitly calling
Mockito.spy. This change has two advantages - first, it's a tad more
elegant, as it saves the duplicate Mockito.spy calls. Second, and more
importantly, it allows injecting the mocked ExternalVmMacsFinder,
MacPoolPerCluster and VmDeviceUtils without an additional boilerplate
code.

Change-Id: Ica69b0e1ce385e7c0ee0b7bdd758edcf94e4fe4b
Signed-off-by: Allon Mureinik <amureini@redhat.com>