core: HotPlugDiskToVmCommandTest @Spy annotation
Spy the command with the @Spy annotation instead of explicitly calling
Mockito.spy in order to set the ground for using injected mocks in
these tests in the future.

Note that VmCommand initializes its osRepository directly as a data
members with a SimpleDependencyInjector, so it must be mocked before the
command is created - i.e., in a @BeforeClass method in a static context.

Change-Id: Iddf8b1ae36e0466dc4b6e121fbac05789834788f
Signed-off-by: Allon Mureinik <amureini@redhat.com>