core: MoveOrCopyDiskCommandTest @Spy annotation
Spy the command with the @Spy annotation instead of explicitly calling
Mockito.spy in order to set the ground for using injected mocks in
these tests in the future.

Note that the MoreOrCopyImageGroupParameters#setOperation method had
to be exposed in order to allow overriding its value after the
parameters object was created.

Change-Id: Icdb834aa28e2d54d0c4dfc397d647a82114f4363
Signed-off-by: Allon Mureinik <amureini@redhat.com>