core: RunVmCommandTest @Spy annotation
Spy the command with the @Spy annotation instead of explicitly calling
Mockito.spy. This change has two advantages - first, it's a tad more
elegant, as it saves the duplicate Mockito.spy calls. Second, and more
importantly, it sets the ground for using injected mocks in these tests
in the future.

Change-Id: Iae9ddd68a57ff72508160ae00a26599033dbcca8
Signed-off-by: Allon Mureinik <amureini@redhat.com>