core: RemoveDiskSnapshotsCommandTest @Spy annotation
Spy the command with the @Spy annotation instead of explicitly calling
Mockito.spy. This change has two advantages - first, it's a tad more
elegant, as it saves the duplicate Mockito.spy calls. Second, and more
importantly, it sets the ground for using injected mocks in these tests
in the future.

Change-Id: I8754a19705c3a2c252f41a91102e26d7393048cf
Signed-off-by: Allon Mureinik <amureini@redhat.com>