core: TryBackToAllSnapshotsOfVmCommandTest @Spy annotation
Spy the command with the @Spy annotation instead of explicitly calling
Mockito.spy. This change has two advantages - first, it's a tad more
elegant, as it saves the duplicate Mockito.spy calls. Second, and more
importantly, it sets the ground for using injected mocks in these tests
in the future.

Note: A public setDstSnapshotId was added to
TryBackToAllSnapshotsOfVmParameters in order to facilitate this
change.

Change-Id: I16a39f9fb3cb09cad18a926eda8c817c24005cca
Signed-off-by: Allon Mureinik <amureini@redhat.com>