core: RestoreAllSnapshotsCommandTest @Spy annotation
Spy the command with the @Spy annotation instead of explicitly calling
Mockito.spy. This change has two advantages - first, it's a tad more
elegant, as it saves the duplicate Mockito.spy calls. Second, and more
importantly, it sets the ground for using injected mocks in these tests
in the future.

Change-Id: If72ef46473b9839091c88c80b5947be3f2aabc1d
Signed-off-by: Allon Mureinik <amureini@redhat.com>