core: Pool command tests @Spy annotation
Spy the commands with the @Spy annotation instead of explicitly calling
Mockito.spy. This change has two advantages - first, it's a tad more
elegant, as it saves the duplicate Mockito.spy calls. Second, and more
importantly, it sets the ground for using injected mocks in these tests
in the future.

Change-Id: I8913ed27aeadba957385055122368c718d284077
Signed-off-by: Allon Mureinik <amureini@redhat.com>