core: RemoveSnapshotCommandTest @Spy annotation
Spy the command with the @Spy annotation instead of explicitly calling
Mockito.spy. This change has two advantages - first, it's a tad more
elegant, as it saves the duplicate Mockito.spy calls. Second, and more
importantly, it sets the ground for using injected mocks in these tests
in the future.

Change-Id: Ic9f18e6a3b0836045724ad728263141197345ca1
Signed-off-by: Allon Mureinik <amureini@redhat.com>