core: ProcessOvfUpdateForStoragePoolCommandTest @Spy annotation
Spy the command with the @Spy annotation instead of explicitly calling
Mockito.spy in order to set the ground for using injected mocks in
these tests in the future.

Note that the command initializes its vmDeviceUtils directly as a data
members with an  Injector, so it must be mocked before the command is
created - i.e., in a @BeforeClass method in a static context.

Change-Id: I6942c896568322f95e2283bbf41f71c2ee67b044
Signed-off-by: Allon Mureinik <amureini@redhat.com>