core: storage connection command tests @Spy annotation
Spy the command with the @Spy annotation instead of explicitly calling
Mockito.spy. This change has two advantages - first, it's a tad more
elegant, as it saves the duplicate Mockito.spy calls. Second, and more
importantly, it sets the ground for using injected mocks in these tests
in the future.

Change-Id: I2744292707142216c1906c70c5ea5225faf584c4
Signed-off-by: Allon Mureinik <amureini@redhat.com>