restapi: Update to model 4.0.35 and metamodel 1.0.20
The more relevant changes in the new version of the model are the
following:

* Fix the type of the `ticket` parameter of the `VmService.ticket`
  method.

* Fix the type of the `authentication_method` attribute of the `Ssh`
  type.

* Rename the `AuthenticationMethod` enum type to `SshAuthenticationMethod`.

* Fix the name of the `exclusive` parameter o the `TemplateService.export`
  method.

* Add the missing `cluster` parameter to the `OpenstackImageService.import`
  method.

The more relevant change in the new version of the metamodel is that it
supports custom attributes in the generated AsciiDoc, and we actually
use the "product-name" and "engine-name" attributes to support branding
of the documentation. That means that we need to add "product.name" and
"engine.name" properties to the maven build, so that we will be able to
produce an engine with branded documentation using those properties. For
example, Red Hat could use this mechanism to use "RHV" instead of
"oVirt" and "RHV Manager" instead of "oVirt Engine":

  $ mvn clean install \
  -Dproduct.name='RHV' \
  -Dengine.name='RHV Manager'

Change-Id: I13f46b544f8127438746ca4314e6d6fbf45a103a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>