core: EditAffinityGroupCommandTest @Spy annotation
Spy the command with the @Spy annotation instead of explicitly calling
Mockito.spy in order to use injected mocks instead of explicitly
spying their return.

Note that as the order of execution between @Mock and @Spy can't be
relied upon the parameters objects was converted from a mock to a real
instance. As this object is a simple data-POJO, there is no real reason
to mock it anyway, and doing so just adds complexity to the test.

Change-Id: Ia11e4739a6ca7277fa04b1a21c321d3629fd4cf6
Signed-off-by: Allon Mureinik <amureini@redhat.com>