core: fix LockingInvocationHandler so that it does not throw UndeclaredThrowableException
when using java.reflect.Proxy, proxied instance might fail with
runtime exception. In that case, if InvocationHandler contains
simple:

return method.invoke(instance, args);

such runtime exception will be 'upgraded' to controlled
InvocationTargetException. If caller is not aware of proxy being used
(which is kinda a point in using them), and he will not declare it in
throws clause (which he shouldn't do anyways) then this controlled
exception will be wrapped inside of UndeclaredThrowableException
runtime exception. This is totally not desirable.

Change-Id: I92e226147710666559781a42e3c25fc619dd727e
Bug-Url: https://bugzilla.redhat.com/1378045
Signed-off-by: Martin Mucha <mmucha@redhat.com>