core: ImportVMFromConfigurationCommandTest command initialization
Moved the initialization of the tested command to the @Before method
in preparation of moving it to be initialized by a @Spy annotation.

Note that the SimpleDependencyInjector is used to initialize the
command's data members so it must be mocked before the command instance
is created. Hence, this patch moves its handling to a @BeforeClass
method in order to be on the safe side.

Change-Id: I2cc73afe0167d1618e5ff76fae1db2d3a3c563b7
Signed-off-by: Allon Mureinik <amureini@redhat.com>