core: CreateGlusterVolumeSnapshotCommand#init
Move the initialization logic from the constructor to the init method,
as per the project's standard.

The init() method is called as part of the command construction
framework, and since it's done in the postConstruct phase, it's safe to
use DAO calls there, allowing the test to be cleaned up in the following
patch.

Change-Id: I803ad6edb87d5096e1de97c2afc8223bf86e7e53
Signed-off-by: Allon Mureinik <amureini@redhat.com>