core: GlusterVolumeSnapshotCommandBase#init
Move the initialization logic of this class and its descendants from
the constructor to the init method, as per the project's standard.

The init() method is called as part of the command construction
framework, and since it's done in the postConstruct phase, it's safe to
use DAO calls there, allowing the test to be cleaned up in the following
patch.

Change-Id: I73eba6d497cd802842bee804f442e757e26a5ec4
Signed-off-by: Allon Mureinik <amureini@redhat.com>