core: DeleteAllGlusterVolumeSnapshotsCommand#init
Move the initialization logic from the constructor to the init method,
as per the project's standard.

The init() method is called as part of the command construction
framework, and since it's done in the postConstruct phase, it's safe to
use DAO calls there, allowing the test to be cleaned up in the following
patch.

Change-Id: I34f1e7c273ae509aad2e670dde38dfe7a666c712
Signed-off-by: Allon Mureinik <amureini@redhat.com>