core: Make CommandHelper.canDoAction accept CommandContext
Currently, the CommandHelper gets ExecutionContext, which lacks the
session ID of the caller command. Building the command without a
session ID, resulting in having the command without a user in its context.

This can cause an unexpected behaviour in some cases, where command
validations rely on the currnt user.

When developers uses CommandHelper.canDoAction, they don't need to worry
about that- they should just clone the command's context and run
CommandHelper.canDoAction.

This patch changes the signature of CommandHelper.canDoAction, to get a
CommandContext.

Change-Id: I48e213de4a15480047952a7911de93b4076c040b
Related-To: https://bugzilla.redhat.com/64378
Signed-off-by: Amit Aviram <aaviram@redhat.com>