core: ImportVmCommandTest single command instance
Rework the test to use a single command instance, annotated with @Spy
and @InjectMocks in order to reduce code duplication and allow easier
injections.

Note that the OsRepository is used to initialize a data member in
VmCommand,  so it must be mocked in a static context, before the
command instance is created.

Change-Id: Ieced27c9f6ff5b60ece9a8927133b9b9833804b1
Signed-off-by: Allon Mureinik <amureini@redhat.com>