core: UpdateVmCommandTest single command instance
Rework the test to use a single command instance, annotated with @Spy
and @InjectMocks in order to reduce code duplication and allow easier
injections.

Note that the OsRepository is used to initialize a data member in
VmCommand, so it must be mocked in a static context, before the
command instance is created.

Change-Id: Id91070d1cde2ea8a5d4a8b00ec49d57237786fbb
Signed-off-by: Allon Mureinik <amureini@redhat.com>