core: RemoveQuotaCommandTest single command instance
Rework the test to use a single command instance, annotated with @Spy
and @InjectMocks in order to reduce code duplication and allow easier
injections.

Instead of using different Quota objects, the test setup initializes a
single Quota data member which specific tests can continue and
specialize (e.g., by setting it to be the default quota).

Change-Id: I97a048f6d4749c915a735cb904a08c15f78347b3
Signed-off-by: Allon Mureinik <amureini@redhat.com>