core: AddVdsCommandTest spying
Mocking an object and calling doRetrunRealMethod() on most of its
methods is essentially spying it, but with the wrong default (a mocked
result instead of the real one).

Since the command is the actual object we want to test, it should NOT be
mocked, and only the relevant parts (e.g., where it interacts with
external resources) should be spied away.

Change-Id: I804128d81482a83c26b6da595c6b7162f9492a4c
Signed-off-by: Allon Mureinik <amureini@redhat.com>