core: Fix "succeeded" flag when retriving VM certificate
Currently the command that retrieves the subject of the certificate of a
virtual machine returns correctly the value, but doesn't set the
"succeeded" flag to "true". As a result the API considers that operation
failed, and doesn't populate the corresponding attributes of the VM
object. This patch fixes that issue.

Change-Id: Ibff406a5746fdecb6604ab2995ff468ad9ad74fc
Bug-Url: https://bugzilla.redhat.com/1378933
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>