webadmin: Don't clean upload model on dialog closing.
Patch I363d06219af83bd21e6384840bf9a969331a9959 introduced a better
handling of model objects, which makes the dialogs in the system to
clean up their related models.

In UploadImageModel, the model is still alive as it is used to upload
the user's file- so it shouldn't be cleaned while the upload dialog is
closed.

This patch removes the cleaning of the model in UploadImagePopupView.

Change-Id: I91b40d5a78d190c16d78e950c892638ea839580e
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1368101
Signed-off-by: Amit Aviram <aaviram@redhat.com>
Signed-off-by: Alexander Wels <awels@redhat.com>