core: adding vg_metadata_device to storage domain static
A support to report the vg metadata device of a block storage
domain was added to vdsm.
Performing operations on the vg metadata device might be hazardous,
this patch adds the option to persist the vg metadata device
in order to prevent from the users submitting such requests before
failing.

Change-Id: I5d9e9de2cd77a2bc000636f5f9db41cb2fb6ca80
Signed-off-by: Liron Aravot <laravot@redhat.com>