core: conversion of most of vm dictionary to domxml
This patch introduces the LibvirtVmXmlBuilder that will ideally
convert all the content in the VM dictionary that is being sent
to VDSM into a complete domxml.

Although it is not yet complete and not used in any way, we
merge it now so future changes in the VM dictionary would also
be done in this class.

Change-Id: Iba3699e7d3f0853fc178dd669b6e9fd13f24a180
Signed-off-by: Arik Hadas <ahadas@redhat.com>