core: Remove VmDeviceUtils from DiskValidator constructor
Patch 6fa804a1 moved VmDeviceUtils (VDU) to CDI, since VDU is used in
DiskValidator as a part of the change VDU was passed to the validator in
the constructor, this approach is wrong since a validator should not get
in the ctor anything but the validated entity let alone various utils.
This patch removes VDU from the ctor of DiskValidator and uses
Injector::get to get an instance of VDU in the single method which uses it.
Also a redundant usage of injectorRule.bind of VDU was removed.

Change-Id: I77e3e8c409a8e488fd4e349995c4b4bd0c2091d3
Signed-off-by: Tal Nisan <tnisan@redhat.com>