aaa: Consolidate SSO session validation requests
Engine makes multiple queries to SSO
to validate each exisitng session. This
could cause performance issue when there
are thousands of sessions. Combining all
requests into one request to fetch all
the statuses in one request can resolve
the performance issue.

When SSO clears an engine session associated
with a token by invoking TerminateSessionsForToken
SSO_ACCESS_TOKEN_PARAMETER value in the
map is null. This can lead to problems when
equality is checked with the contents of
the map for the token.

Change-Id: Ie63dd4e85ac4e4b25cd2884d07d84483dc8add17
Bug-Url: https://bugzilla.redhat.com/1379805
Signed-off-by: Ravi Nori <rnori@redhat.com>