core: AddDiskCommandTest: rm redundant VM mocking
AddDisk deals primarily with adding disks (unsurprisingly), although
it has a small overlap with VMs as it allows to add and attach a disk
in a single action. Most of the test cases, though, do not relate to
VMs, and mocking a VM into them just bloats the code and confuses the
reader.

This patch reduces the VM mocking to where its actually needed in order
to make the test more concise and easier to read.

Change-Id: I0984bd695054dd5143cfa3bd7458a3fb26a8ac78
Signed-off-by: Allon Mureinik <amureini@redhat.com>