hosted-engine: add a DB cleaner utility
Hosted-engine auto-import procedure is not going to work
if the engine DB already contains  hosted-engine entities and,
on the other side, they cannot be removed from the
engine to avoid breaking a working environment.
This was preventing the user from migration an hosted-engine env
from one storage device to another via backup/restore.

Change-Id: Id61ae0b05a75018ded532d7a0c38c15b4b885803
Bug-Url: https://bugzilla.redhat.com/1240466
Bug-Url: https://bugzilla.redhat.com/1235200
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>