core: Remove redundant command spying
@Spying a command and then mocking behavior with
do[Return|Answer|Nothing] is a common practice in BLL's tests which
originated from the need to mock DAOs' behavior.

After the recent changes, the commands' DAOs are injected, so in many
scenarios using @Spy is redundant, and @InjectMocks is sufficient to
introduce the required dependencies.

This patch removes these redundant spying by doing the following:
- Remove the @Spy annotation where possible
- Using straight-forward setter methods instead of spying the behavior
  of getters with doReturn. This has an added bonus of testing another
  method of the command, albeit a simple one.

Change-Id: I828e0f55cec9ccbe873636ca40afb8bbbcf1daf3
Signed-off-by: Allon Mureinik <amureini@redhat.com>