core: QuotaDaoTest: Remove Config dependency
The QuotaDaoTest tests performing database CRUD operations on Quota
objects. The actual values set on the Quota object are
inconsequential for this test, and there's no reason to query them from
Config values in the database.

This patch removes the quota dependency and replaces the Config calls
with the actual values that would have been returned from the test
database.

Change-Id: If5a13d9675f0b49d663fa6a1e46e75c49c5b3bb9
Signed-off-by: Allon Mureinik <amureini@redhat.com>