core: DBConfigUtils dependent values with defaults
The DBConfigUtils implementation for dependent config values searched
for the dependent value in the cache, causing it to fail if that value
was not present in the database.

This patch changes the explicit retrieval from the cache with a
getValue call, thus allowing dependent values to use the default value
if there is no value in the database.

Change-Id: I2f822255ddb91974ea238c1e85747b36d68b0bf1
Signed-off-by: Allon Mureinik <amureini@redhat.com>