core: Remove VDS#getCertifaicateSubject
Business entities are shared between the back backend and the
frontend, and as such should be simple data POJO and not contain any
logic.

This patch removes the complicated VDS#getCertificateSubject and changed
the few places that used it to use the CertificateSubjectHelper
directly.

This decouples the VDS[Static] class from the config value and allows
for better layering of the project's logic.

Change-Id: I219df6ed8b40b93645a38b49883b927ac1e9fd2c
Signed-off-by: Allon Mureinik <amureini@redhat.com>