core: MockConfigRule default values
Enhance MockConfigRule to return the default value of a config value if
its not explicitly mocked.

Note that now the MockConfigRule depends on production code from the
utils module, so modules that do not depend on it must have it added (in
the test scope)

This change has four main advantages:
1. Ease of use: With this change, not every config value has to be
   mocked, only the ones the developer is explicitly interested in.
2. Better simulation: With this change, tests with MockConfigRule act
   more like production code - if a config value is missing in their
   [mocked] environment, the default value is used.
3. Ease of maintenance: With this change, adding a new config value
   dependency in the production code won't break all the tests with a
   NullPointerException. Instead, the tests will silently use the
   default value.
4. Ease of reading: With this change, the only MockCongfigDescriptors
   that need to be used are for exceptional values that differ from
   the defaults. This way, these values draw the reader's attention
   to their special nature and allow him to focus on them instead of
   loosing them between all the values that just mock their defaults.

NOTE:
The SearchQueryTest relies on the old behavior. In order to not break
it, nulls were explicitly mocked where expected by that old test.

Change-Id: I32c99fd9cba88d2f58db4878565251a30a0bf125
Signed-off-by: Allon Mureinik <amureini@redhat.com>