core: UpgradeOvirtNodeInternalCommandTest: Clear OVirtNodeInfo
OVirtNodeInfo stores a Config value in a static context, so relying
on it means this test is susceptible to outside interference if other
tests use it but neglect to clean it up, causing unexpected test
failures.

This patch add a safe-guard to the test by clearing the instance
before the test is run.

Change-Id: I451fd2dfae005e23c49dcd829facee66e6531f1e
Signed-off-by: Allon Mureinik <amureini@redhat.com>