core: ManageGlusterServiceCommand: inline members
Keeping members in the command that are just duplicates of the values
form the parameters objects leads to data duplicity and potentially
nasty bugs.
This patch clears up the duplicity by using the values directly from the
parameters. As a side bonus, this also makes the code cleaner and
testing much easier.

Change-Id: Ice26853743c914aa7bd448568fc5d808d8fa291f
Signed-off-by: Allon Mureinik <amureini@redhat.com>