restapi: Don't populate NetworkAttachment.network.name
Currently when the list of network attachments of a host is retrieved we
return a link to the network that includes also the network name:

  GET /ovirt-engine/api/hosts/123/networkattachments

  <network_attachments>
    <network_attachment id="..." href="...">
      ...
      <network id="..." href="...>
        <name>mynetwork</name>  <-- This should't appear
      </network>
    </network_attachment>
    ...
  </network_attachments>

This is not correct, as output links should not have any content other
than the "id" and "href" attributes. This patch fixes that issue.

Change-Id: Ie0eb5ba57a03f0e89c844d22969015ddf67a4134
Bug-Url: https://bugzilla.redhat.com/1382341
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 27209c687d3f30228ea33c20a80b39bb46d6d7c3)