core: fix check for whether vm is starting up
Previously the check relied on the result of getLastVdsRunOn to see if
the VM is starting up since preferred hosts are supposed to be used
only during VM startup. But VmDynamic#lastVdsRunOn never contains
the real host that the VM ran on when getting to this method (it is
always null so this check is passed also for migrating VMs).

Therefore, this check is changed.

Change-Id: I02832a76d94f7a02d6f2b82970461d14a628ed74
Signed-off-by: Arik Hadas <ahadas@redhat.com>