core: simplify fullList on json-rpc at the expense of xml-rpc
FullListVDSCommand used to convert the given VM IDs to an array because the
XML-RPC based verb expects to get an array. Now that JSON-RPC is the default
and more common protocol, it is time to change this.

So now FullListVDSCommand passes a list that is used as-is with JSON-RPC,
and with XML-RPC it is converted to an array as before.

Change-Id: I3a2e60f2fa2ecbe1b78f588401626d15ed2bb9fc
Signed-off-by: Arik Hadas <ahadas@redhat.com>