core: report SwitchType also on bonds
• all networks get same switch type, so it need not be specified per
HostNetwork. It should be sufficient to have it specified only once in
paramaters class.

• onces that's done, we need not to update CreateOrUpdateBond class
with SwitchType just as HostNetwork used to contain it
(which would be totally incorrect(*)), and use SwitchType from
parameters class.

(*) it's problematic to share TO to pass data between more than 2 app
layers. It can work, until all its data makes sense everywhere, but
here we need only to pass switch type of CreateOrUpdateBond from bll
to vdsm. It's totally meaningless for UI and thus it must not be added
into this class.

Change-Id: Id5a8d2ebc70eb50a73d9c379beb0ea18e5a50ef1
Signed-off-by: Martin Mucha <mmucha@redhat.com>