aaa: Assign unique id to each logged in user
On user login if the user is not found in
the database a default user id was assigned
to the all users (all zeros). Now when two
users logged in to the system they got the
union of permissions for each user.

This can be resolved by assigning a unique id
to each logged in user if the user is not
found in the database

Change-Id: Ie79a1d0b9b5da4fa338e166510b678212e8e137f
Bug-Url: https://bugzilla.redhat.com/1380780
Signed-off-by: Ravi Nori <rnori@redhat.com>