core: Rm ChangeVDSClusterCommand#targetClusterCompatibilityVersion
Calculating targetClusterCompatibilityVersion requires access to the
ClusterDao, and doing so in a data member means that DAO cannot be
injected. Since this member is only used in one place anyway, it makes
very little sense holding it as a member, so this patch inlines the
call.

Doing so has some advantages:
1. As mentioned above, it allows injecting the DAO
2. Performance: The command may fail its validation before the
   targetCluster is required. In this case, a database call is saved.

Change-Id: Ie01547f3b3e3b87dc6addd313e7d4f8ef418ebbf
Signed-off-by: Allon Mureinik <amureini@redhat.com>