engine: Adding activeSlave to Bond
To prepare exposing the information of the active slave in a "bonded"
interface, the information is added to the data model.

This patch includes the extension the extension of the business entity
class Bond and the corresponding mapping in the data base and an update
of the fixtures of dao-test.

Add missing columns of table vds_interface to fixtures of dao-test.
The order of the columns in the fixture is now consistent to the one in
the data base table.

Change-Id: Ie477f737a3f3b77f4611a90cb53005f529af32cd
Bug-Url: https://bugzilla.redhat.com/1038550
Signed-off-by: Dominik Holler <dholler@redhat.com>