webadmin: Add RpmVersion#hashCode
Having a class implement equals(Object) but not implement hashCode() may
cause hard to debug bugs when the object used in a hashed context (e.g.,
placed in a HashSet or used as a key in a HashMap).
Even if we do not currently use RpmVersion in such contexts,
implementing it is a good defensive practice to avoid potential bugs in
the future.

Change-Id: I1aa6ae1b9a1f9d08d586cc156c5b68b364e4a90b
Signed-off-by: Allon Mureinik <amureini@redhat.com>