webadmin: FormatterJava case scopes
Removed the anonymous blocks in the switch-case statement as per the
project's checkstyle requirement.
Note that in order to achieve this the variables i, flags and sep were
extracted outside of the switch block. In the one place where i was
defined as a long the expression was inlined to avoid a compilation
error.

Change-Id: Ie8e1e0dd92531158759ebc9f62e5f8d45aba06af
Signed-off-by: Allon Mureinik <amureini@redhat.com>