webadmin: Present the name of the active slave in a "bonded" interface
The information of the active slave in a "bonded" interface is presented
to the user.

The already existing tooltip of the bond icon in the BondPanel currently
shows properties, like the status, of the bond.

This patch extends the tooltip by an additional line, which contains the
name of the active slave interface. This is done in the BondPanel,
because the active slave is handled as a read-only property of the bond.

Change-Id: I8ac5e308eba8774388372a60c4fccaef99f04385
Bug-Url: https://bugzilla.redhat.com/1038550
Signed-off-by: Dominik Holler <dholler@redhat.com>