core: Fix move and copy disk commands
Change I4bddbcc changed the initialization of the storage pool in
BaseImagesCommand and in case the storage pool from the parameters is not
null it will be taken from the parameters, however this change ignored
the case in which the storage pool from the parameters is an empty and it
that case skipping the part of getting the storage pool id from the image
fetched from the DB, this patch adds the check for the empty guid.

Since the original patch also changed the static mocking of the commands
to do nothing when invoking initContainerDetails(params) due to the fact
it issues a DAO call from the ctor the tests are breaking since the DAO
mocks are not populated upon the ctor call to initContainerDetails was
moved to the init() method of the command where it belonged to in the first
place as issuing DAO calls from the ctor is a bad practice.

Bug-Url: https://bugzilla.redhat.com/1385533
Change-Id: Ibe620a55fbfd509689a081e5c658a9e25819dd26
Signed-off-by: Tal Nisan <tnisan@redhat.com>